fix: build breaks and correctness bugs in engine sync/spark/fuel paths
* Fix build breaks in sync/spark paths to restore host-test harness

- ckp.cpp: hoist is_forward_rotation_coherent() and its state above
  process_gap_event(), which uses them (was a use-before-declaration
  introduced by the BUG-12 patch; the sync module did not compile).
- timer.cpp: the TIM1/ETB PWM functions banged registers that are not
  defined in regs.h and sat in the EMS_HOST_TEST branch; replace with
  empty host stubs matching the tim3/tim4 pattern so the harness builds.
- ign_calc.cpp: add a weak ems::engine::etb_get_idle_spark_trim() default
  returning 0 so the idle-spark path is self-contained when the ETB unit
  (global-namespace, not in the engine build) is absent.


* Fix x-tau int32 overflow and map estimator faulted-sensor passthrough

- transient_fuel.cpp / xtau_autocalib.cpp: numerator_q8 * 256 overflows
  int32_t for pulse widths near kMaxPwUs (>~32.7 ms). Use a 64-bit
  intermediate for the divide; the post-clamp result still fits int32_t.
- map_estimator.cpp: an out-of-range MAP reading set estimator_mode=2 but
  still fed the raw (railed/0) value into the complementary filter. Force
  the sensor gain to 0 when the reading is invalid so the output falls back
  to the pure model.


* Fix pipeline scope/include bugs in main_stm32.cpp integration layer

- snap was referenced in the 500ms flash block and the per-iteration
  adaptive-flush block but only declared inside the 2/20/100ms blocks
  (undeclared identifier). Take a fresh ckp_snapshot() in each scope.
- Add missing #include "engine/constants.h" for kFlashWriteSafeRpmX10.

Remaining API-drift in this file (ETB pedal source, x-tau lambda/o2
source) is documented in the PR for direction.


* Route 2ms x-tau autocalib lambda from CAN WBO2

The 2ms transient-fuel x-tau autocalibration read sensors.o2_valid and
sensors.lambda_raw_x1000, which no longer exist (O2/lambda was moved to
WBO2-over-CAN only). Mirror the 100ms STFT path: gate on
can_stack_wbo2_fresh() and read can_stack_lambda_milli_safe(), clamped to
the same lambda window. Leaves only the ETB accelerator-pedal source
(no APP field in SensorData) as a documented integration gap.


* Wire ETB pedal source (AN1) and complete TIM1 HAL definitions

Item 1 - ETB pedal source:
- main_stm32.cpp read sensors.app_pct_x10, which does not exist. There is
  no dedicated accelerator-pedal (APP) signal in the chain, so source the
  torque-manager pedal from the AN1 expansion analog input (12-bit raw ->
  percent). APP is intentionally distinct from tps_pct_x10 (throttle-plate
  feedback) as required for drive-by-wire.
- Fix etb_control.h include to the repo convention (hal/etb_driver.h) so the
  integration layer compiles under -I./src.

Item 2 - TIM1 HAL:
- regs.h: add TIM1 (advanced timer, APB2, base 0x40012C00) register
  accessors + RCR offset and the missing TIM_EGR_UG / TIM_BDTR_MOE /
  TIM_CCER_CC1NE / RCC_APB2ENR_TIM1EN bit macros.
- timer.cpp: restore the real ETB PWM implementation in the hardware branch,
  correcting the clock-enable bus from APB4 to APB2 (TIM1 is on APB2); keep
  empty host stubs.

Validated: host-test harness passes; timer.cpp and main_stm32.cpp both
compile clean under -DTARGET_STM32H562 -I./src. Full firmware link still
needs the ETB/torque .cpp added to FIRMWARE_SRC, and etb_driver.cpp ported
off CMSIS (stm32h5xx.h) to this repo's regs.h HAL.


---------

+++ src/engine/xtau_autocalib.cpp
@@ -278,7 +278,10 @@ uint32_t transient_fuel_xtau_with_autocalib(uint32_t fuel_pw_us,
     }
     
     const int32_t dry_fraction_q8 = 256 - static_cast<int32_t>(x_q8);
-    int32_t injected_q8 = (numerator_q8 * 256) / dry_fraction_q8;
+    // numerator_q8 chega a ~25.6M (kMaxPwUs<<8); ×256 estoura int32_t.
+    // Intermediário de 64 bits evita overflow; o resultado pós-clamp cabe em int32_t.
+    int32_t injected_q8 = static_cast<int32_t>(
+        (static_cast<int64_t>(numerator_q8) * 256) / dry_fraction_q8);
     const int32_t max_q8 = static_cast<int32_t>(kMaxPwUs << 8u);
     if (injected_q8 > max_q8) {
         injected_q8 = max_q8;
